# docs/swagger.yaml
basePath: /api/v1
definitions:
  main.ActivityItem:
    properties:
      action:
        type: string
      collection:
        type: string
      comment:
        type: string
      id:
        type: string
      ip:
        type: string
      item:
        type: string
      timestamp:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  main.CollectionMetrics:
    properties:
      collections_by_type:
        additionalProperties:
          type: integer
        type: object
      most_active_collections:
        items:
          $ref: '#/definitions/main.CollectionSummary'
        type: array
      recent_collections:
        items:
          $ref: '#/definitions/main.CollectionSummary'
        type: array
      total_collections:
        type: integer
    type: object
  main.CollectionModel:
    properties:
      created_at:
        example: "2023-01-01T10:30:00Z"
        type: string
      description:
        example: Product catalog collection
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: Products
        type: string
      schema:
        example: '{}'
        type: string
      updated_at:
        example: "2023-12-01T10:30:00Z"
        type: string
    type: object
  main.CollectionSummary:
    properties:
      collection:
        type: string
      created_at:
        type: string
      hidden:
        type: boolean
      icon:
        type: string
      item_count:
        type: integer
      note:
        type: string
      singleton:
        type: boolean
    type: object
...
  main.DashboardOverview:
    properties:
      collection_metrics:
        $ref: '#/definitions/main.CollectionMetrics'
      recent_activity:
        items:
          $ref: '#/definitions/main.ActivityItem'
        type: array
      system_health:
        $ref: '#/definitions/main.SystemHealth'
      system_stats:
        $ref: '#/definitions/main.SystemStats'
      user_insights:
        $ref: '#/definitions/main.UserInsights'
    type: object
  main.ErrorResponse:
    properties:
      error:
        example: Invalid request payload
        type: string
    type: object
...
  main.SystemHealth:
    properties:
      database_connected:
        type: boolean
      last_backup:
        type: string
      server_uptime:
        type: string
      version:
        type: string
    type: object
  main.SystemStats:
    properties:
      active_users:
        type: integer
      total_collections:
        type: integer
      total_roles:
        type: integer
      total_sessions:
        type: integer
      total_users:
        type: integer
    type: object
...
  main.UserInsights:
    properties:
      most_active_users:
        items:
          $ref: '#/definitions/main.UserSummary'
        type: array
      new_users_this_month:
        type: integer
      new_users_this_week:
        type: integer
      recent_registrations:
        items:
          $ref: '#/definitions/main.UserSummary'
        type: array
      users_by_role:
        additionalProperties:
          type: integer
        type: object
      users_by_status:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  main.UserSummary:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_access:
        type: string
      last_name:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@gorectus.com
    name: GoRectus API Support
  description: A comprehensive API for managing collections, items, fields, users,
    and roles
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: GoRectus API
  version: "1.0"
paths:
...
  /collections:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all collections in the system
      produces:
      - application/json
      responses:
        "200":
          description: List of collections
          schema:
            items:
              $ref: '#/definitions/main.CollectionModel'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all collections
      tags:
      - collections
    post:
      consumes:
      - application/json
      description: Create a new collection in the system
      parameters:
      - description: Collection data
        in: body
        name: collection
        required: true
        schema:
          $ref: '#/definitions/main.CollectionModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created collection
          schema:
            $ref: '#/definitions/main.CollectionModel'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "409":
          description: Collection already exists
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new collection
      tags:
      - collections
...
  /dashboard:
    get:
      consumes:
      - application/json
      description: Retrieve comprehensive dashboard overview including system stats,
        collection metrics, recent activity, system health, and user insights (user
        insights only available for admins)
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard overview data
          schema:
            additionalProperties:
              $ref: '#/definitions/main.DashboardOverview'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get dashboard overview
      tags:
      - dashboard
  /dashboard/activity:
    get:
      consumes:
      - application/json
      description: Retrieve recent system activity with optional limit parameter
      parameters:
      - description: 'Maximum number of activity items to return (1-100, default:
          20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Recent activity data
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/main.ActivityItem'
              type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get recent activity
      tags:
      - dashboard
  /dashboard/collections:
    get:
      consumes:
      - application/json
      description: Retrieve collection metrics including total collections, distribution
        by type, and activity data
      produces:
      - application/json
      responses:
        "200":
          description: Collection metrics data
          schema:
            additionalProperties:
              $ref: '#/definitions/main.CollectionMetrics'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get collection insights
      tags:
      - dashboard
  /dashboard/stats:
    get:
      consumes:
      - application/json
      description: Retrieve system statistics including user counts, role counts,
        collection counts, and active sessions
      produces:
      - application/json
      responses:
        "200":
          description: System statistics
          schema:
            additionalProperties:
              $ref: '#/definitions/main.SystemStats'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get system statistics
      tags:
      - dashboard
  /dashboard/users:
    get:
      consumes:
      - application/json
      description: Retrieve user insights including user distribution by status and
        role, new user statistics, and user activity data (Admin only)
      produces:
      - application/json
      responses:
        "200":
          description: User insights data
          schema:
            additionalProperties:
              $ref: '#/definitions/main.UserInsights'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - Admin access required
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user insights
      tags:
      - dashboard
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"